Core refactoring
- move core logic to a separate directory
- move the generic logic to a
- move block logic to a separate file
- create structure for sdk-specific logic
