+++ services/core/methods/controllers/accounts.js
@@ -18,8 +18,8 @@ const { HTTP, Utils, Logger } = require('lisk-service-framework');
 const { StatusCodes: { NOT_FOUND } } = HTTP;
 const { isEmptyArray, isEmptyObject } = Utils.Data;
 
-const CoreService = require('../../shared/core.js');
-const config = require('../../config.js');
+const CoreService = require('../../shared/core');
+const config = require('../../config');
 
 const logger = Logger();
