+++ services/core/methods/controllers/status.js
@@ -17,7 +17,7 @@ const { HTTP } = require('lisk-service-framework');
 
 const httpStatus = HTTP.StatusCodes;
 
-const CoreService = require('../../shared/core.js');
+const CoreService = require('../../shared/core');
 const packageJson = require('../../package.json');
 
 const getStatus = async () => {
