+++ services/core/shared/core/sdk_v4/index.js
@@ -1,6 +1,6 @@
 /*
  * LiskHQ/lisk-service
- * Copyright © 2019 Lisk Foundation
+ * Copyright © 2020 Lisk Foundation
  *
  * See the LICENSE file at the top-level directory of this distribution
  * for licensing information.
@@ -13,22 +13,4 @@
  * Removal or modification of this copyright notice is prohibited.
  *
  */
-const CoreService = require('../../shared/core');
-
-const getNetworkStatus = async () => {
-	const status = await CoreService.getNetworkStatus();
-	const constants = await CoreService.getNetworkConstants();
-	const result = {};
-	result.status = status.data;
-	result.constants = constants.data;
-
-	return {
-		data: result,
-		meta: {},
-	};
-};
-
-
-module.exports = {
-	getNetworkStatus,
-};
+module.exports = require('../sdk_v2');
